warning indicators

+++ src/fakeroot/usr/share/cockpit/ceph-deploy/ceph-deploy.js
@@ -3386,32 +3386,51 @@ function check_for_parameter_change(param_json_msg){
         if(param_group["global"].length > 0){
           //there are variables that pertain to a specific group that may have been modified
           param_group["global"].forEach((global_option) =>{
-            console.log("here 1");
             if(
               old_params["options"].hasOwnProperty(global_option.option_name) && 
               new_params["options"].hasOwnProperty(global_option.option_name) &&
               old_params["options"][global_option.option_name] != new_params["options"][global_option.option_name]
             ){
-              console.log("here 2");
               let deploy_state = localStorage.getItem("ceph_deploy_state");
               let deploy_state_json = JSON.parse(deploy_state);
               Object.entries(deploy_state_json).forEach(([deploy_step_id,state_vars]) => {
-                console.log("here 3");
                 if(state_vars["lock_state"] === "complete" && g_role_to_deploy_step_lut[role_name].includes(deploy_step_id)){
                   if(!deploy_state_json[deploy_step_id].hasOwnProperty("warning_vars")){
                     deploy_state_json[deploy_step_id]["warning_vars"] = [];
                   }
-                  console.log("here 4");
-                  let warning_var = {
-                    "var_name": global_option.option_name,
-                    "original_value": old_params["options"][global_option.option_name],
-                    "current_value": new_params["options"][global_option.option_name],
-                    "original_time_stamp": old_params["time_stamp"]
-                  };
-                  deploy_state_json[deploy_step_id]["warning_vars"].push(warning_var);
-                  deploy_state_json[deploy_step_id]["warning_msg"] = "Warning: variables used to complete " + deploy_step_id + " have been modified. You may need to re-do this step.";
-                  localStorage.setItem("ceph_deploy_state",JSON.stringify(deploy_state_json,null,4));
-                  sync_ceph_deploy_state();
+                  else{
+                    // we already have a warning vars array.
+                    let existing_warning_var = search_jarray_key_value_match(
+                                                  deploy_state_json[deploy_step_id]["warning_vars"],
+                                                  "var_name",
+                                                  global_option.option_name);
+                    if(existing_warning_var){
+                      //warning var was already flagged. check to see if the user changed it back to the
+                      //value that was used during deployment.
+                      if(new_params["options"][global_option.option_name] === existing_warning_var["warning_var"]["original_value"]){
+                        // the user changed it back to the proper value, remove the warning variable.
+                        delete deploy_state_json[deploy_step_id]["warning_vars"][existing_warning_var["index"]];
+                        if(deploy_state_json[deploy_step_id]["warning_vars"].length == 0){
+                          delete deploy_state_json[deploy_step_id]["warning_msg"];
+                        }
+                      }
+                    }else{
+                      //warning var was not already found. make a new one
+                      let warning_var = {
+                        "var_name": global_option.option_name,
+                        "original_value": old_params["options"][global_option.option_name],
+                        "current_value": new_params["options"][global_option.option_name],
+                        "original_time_stamp": old_params["time_stamp"]
+                      };
+                      deploy_state_json[deploy_step_id]["warning_vars"].push(warning_var);
+                      deploy_state_json[deploy_step_id]["warning_msg"] = "Warning: variables used to complete " + 
+                                                                          deploy_step_id + " have been modified.\n" + 
+                                                                          JSON.stringify(deploy_state_json[deploy_step_id]["warning_vars"],null,4) + 
+                                                                          "\nYou may need to re-do this step.";
+                      localStorage.setItem("ceph_deploy_state",JSON.stringify(deploy_state_json,null,4));
+                      sync_ceph_deploy_state();
+                    }
+                  }
                 }
               });
             }
@@ -3430,6 +3449,18 @@ function check_for_parameter_change(param_json_msg){
       }
     }
 
+}
+
+function search_jarray_key_value_match(json_array,key,value){
+  for (var i=0; i < json_array.length; i++) {
+      if (json_array[i].hasOwnProperty(key) && json_array[i][key] === value) {
+          return {
+            "warning_var":json_array[i],
+            "index":i
+          };
+      }
+  }
+  return null;
 }
 
 /**
@@ -3500,7 +3531,8 @@ function show_host_file() {
 
 /**
  * uses the make_hosts helper script to create /usr/share/ceph-ansible/hosts.
- * This will use the core_params.json file to create this.
+ * This will use the core_params.json file to create this. It will also
+ * create the inventory files required for each host in /usr/share/ceph-ansible/host_vars/
  */
 function generate_host_file() {
   var spawn_args = ["/usr/share/cockpit/ceph-deploy/helper_scripts/make_hosts"];
